requirements.txt
paho_mqtt==1.6.1
pycryptodomex==3.18
tinyec==0.4.0
click==8.1.3